+++ StepLang.CLI/Commands/DefaultCommand.cs
@@ -1,4 +1,3 @@
-using Spectre.Console;
 using Spectre.Console.Cli;
 using StepLang.Tooling.CLI;
 using System.ComponentModel;
@@ -11,30 +10,42 @@ internal sealed class DefaultCommand : AsyncCommand<DefaultCommand.Settings>
 {
 	[SuppressMessage("ReSharper", "UnusedAutoPropertyAccessor.Global")]
 	[SuppressMessage("ReSharper", "AutoPropertyCanBeMadeGetOnly.Global")]
-	public sealed class Settings : VisibleGlobalCommandSettings
+	public sealed class Settings : CommandSettings, IGlobalCommandSettings
 	{
-		[CommandArgument(0, "[file]")]
+		public const string DefaultProgramFileName = "program.step";
+
+		[CommandArgument(0, $"[file={DefaultProgramFileName}]")]
 		[Description("The path to a .step-file to run.")]
-		[DefaultValue(null)]
+		[DefaultValue(DefaultProgramFileName)]
 		public string? File { get; init; }
+
+		[CommandOption("-w|--no-warn")]
+		[Description("Don't emit warnings when processing the source code.")]
+		public bool NoWarn { get; init; }
+
+		[CommandOption("--info")]
+		[Description("Print the version and system information. Add the output of this to bug reports.")]
+		public bool Info { get; init; }
+
+		[CommandOption("-v|--version")]
+		[Description("Print the version number.")]
+		public bool Version { get; init; }
+
+		public bool Handled { get; set; }
 	}
 
 	public override async Task<int> ExecuteAsync(CommandContext context, Settings settings)
 	{
-		if (settings.File == null)
-		{
-			if (settings is not { Info: false, Version: false })
-			{
-				return 0;
-			}
+		if (settings is { Handled: true })
+			return 0;
 
-			AnsiConsole.MarkupLine("[red]No file specified.[/]");
-
-			return 1;
-		}
+		var fileToRun = settings.File ?? Settings.DefaultProgramFileName;
 
 		var runCommand = new RunCommand();
 
-		return await runCommand.ExecuteAsync(context, new RunCommand.Settings { File = settings.File });
+		return await runCommand.ExecuteAsync(
+			context,
+			new() { File = fileToRun, NoWarn = settings.NoWarn }
+		);
 	}
 }